fix(infra-test): rename helper.ts -> utils.ts
